orng_lab/requirements.txt
numpy
pandas
scipy
scikit-learn
statsmodels
matplotlib
yfinance
vectorbt
backtrader
torch
jupyter
pyyaml
alpaca-trade-api
seaborn
tk
